s2226 - refactor get_difference from the caom_util module to it's own module to avoid circular dependencies.
