+++ caom2/caom2/caom_util.py
@@ -80,7 +80,6 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 import collections
-import math
 import struct
 import sys
 import uuid
@@ -89,11 +88,8 @@ from datetime import datetime
 import six
 from builtins import bytes, int
 
-from caom2.common import CaomObject
 
-
-__all__ = ['TypedList', 'TypedSet', 'TypedOrderedDict', 'ClassProperty',
-           'get_differences']
+__all__ = ['TypedList', 'TypedSet', 'TypedOrderedDict', 'ClassProperty']
 
 # TODO both these are very bad, implement more sensibly
 IVOA_DATE_FORMAT = "%Y-%m-%dT%H:%M:%S.%f"
@@ -226,204 +222,6 @@ def value_check(value, min_value, max_value, variable, override=None):
                     min_value, variable, max_value, value))
 
     return True
-
-
-def get_differences(expected, actual, parent=None):
-    """
-    Compare two entities. Provide a report if differences exist between the two
-    entities.
-
-    :param expected: What is expected to exist. May be AbstractCaomEntity,
-    TypedOrderedDict, TypedList, or TypedSet.
-    :param actual: What exists. May be AbstractCaomEntity,
-    TypedOrderedDict, TypedList, or TypedSet.
-    :param parent: str for message content.
-    :return: None if the entities are the same, or a text report of the
-    individual differences.
-    """
-    report = []
-
-    if type(expected) != type(actual):
-        report.append(
-            'Types:: expected \'{}\' actual \'{}\''.format(type(expected),
-                                                           type(actual)))
-        return report
-
-    if (isinstance(expected, TypedOrderedDict) or
-        isinstance(expected, TypedList) or
-            isinstance(expected, TypedSet)):
-        temp_report = _get_collection_differences(expected, actual, parent)
-    else:
-        assert isinstance(expected, CaomObject)
-        assert isinstance(actual, CaomObject)
-        if parent:
-            parent = '{}.{}'.format(parent, expected.__class__.__name__)
-        else:
-            parent = expected.__class__.__name__
-        temp_report = _get_object_differences(expected, actual, parent)
-
-    if temp_report:
-        report.extend(temp_report)
-
-    return report if len(report) > 0 else None
-
-
-def _get_object_differences(expected, actual, parent):
-    """Reports on the differences between both attributes and
-    their values for object differences."""
-    report = []
-    expected_dict, expected_decompose = _get_dict(expected)
-    actual_dict, actual_decompose = _get_dict(actual)
-
-    if expected_dict != actual_dict:
-        temp_report = _get_dict_differences(expected_dict, actual_dict, parent)
-        if temp_report:
-            report.extend(temp_report)
-
-    for expected_key, expected_value in expected_decompose.items():
-        if expected_key in actual_decompose:
-            actual_value = actual_decompose[expected_key]
-            label = '{}.{}'.format(parent, expected_key)
-            temp_report = get_differences(expected_value, actual_value, label)
-            actual_decompose.pop(expected_key)
-            if temp_report:
-                report.extend(temp_report)
-        else:
-            report.append('Member:: {}.{}: missing from {}'.format(
-                parent, expected_key, actual.__class__))
-
-    for actual_key in actual_decompose.items():
-        report.append('Member:: {}.{}: missing from {}'.format(
-            parent, actual_key, expected.__class__))
-
-    return report if len(report) > 0 else None
-
-
-def _get_collection_differences(expected, actual, parent):
-    """Reports on the differences between two collections. Ignores collection
-    ordering."""
-    report = []
-    if len(expected) != len(actual):
-        report.append(
-            'Collection:: {}: length of expected {} != length of actual {}'.format(
-                parent, len(expected), len(actual)))
-
-    if isinstance(actual, TypedList) or isinstance(actual, TypedSet):
-        temp_report = _get_sequence_differences(expected, actual, parent)
-    else:
-        temp_report = _get_mapping_differences(expected, actual, parent)
-
-    if temp_report:
-        report.extend(temp_report)
-
-    return report if len(report) > 0 else None
-
-
-def _get_mapping_differences(expected, actual, parent):
-    report = []
-    actual_keys_copy = list(actual.keys())
-
-    for expected_key, expected_value in expected.items():
-        label = '{}[\'{}\']'.format(parent, expected_key)
-        if expected_key in actual.keys():
-            actual_value = actual[expected_key]
-            temp_report = get_differences(expected_value,
-                                          actual_value, label)
-            actual_keys_copy.remove(expected_key)
-            if temp_report:
-                report.extend(temp_report)
-        else:
-            report.append('Map:: {} not in actual.'.format(label))
-
-    for key in actual_keys_copy:
-        label = '{}[\'{}\']'.format(parent, key)
-        report.append('Map:: actual {} not in expected.'.format(
-            label))
-
-    return report if len(report) > 0 else None
-
-
-def _get_sequence_differences(expected, actual, parent):
-    report = []
-    actual_copy = list(actual)
-
-    for expected_key, expected_value in enumerate(expected):
-        label = '{}[\'{}\']'.format(parent, expected_key)
-        try:
-            actual_value = actual[expected_key]
-            temp_report = get_differences(expected_value,
-                                          actual_value, label)
-            actual_copy.remove(actual_value)
-            if temp_report:
-                report.extend(temp_report)
-        except:
-            report.append('Sequence:: {} not in actual.'.format(label))
-
-    for key, value in enumerate(actual_copy):
-        label = '{}[\'{}\']'.format(parent, key)
-        report.append('Sequence:: actual {} not in expected.'.format(
-            label))
-
-    return report if len(report) > 0 else None
-
-
-def _get_dict_differences(expected, actual, parent):
-    """Reports on how two dictionaries are different."""
-    report = []
-    for expected_key, expected_value in expected.items():
-        if expected_key in actual:
-            actual_value = actual[expected_key]
-            if _not_equal(expected_value, actual_value):
-                report.append(
-                    'Value:: {}.{}: expected {} actual {}'.format(
-                        parent,
-                        expected_key,
-                        expected_value,
-                        actual_value))
-            actual.pop(expected_key)
-        else:
-            report.append(
-                'Member:: {}.{}: expected missing from actual.'.format(
-                    parent, expected_key))
-
-    for key in actual.items():
-        report.append(
-            'Member:: {}.{}: actual missing from expected.'.format(parent,
-                                                                   key))
-
-    return report if len(report) > 0 else None
-
-
-def _not_equal(rhs, lhs):
-    """Handling for not-quite-equals float comparisons."""
-    if isinstance(rhs, float) and isinstance(lhs, float):
-        result = math.isclose(rhs, lhs, rel_tol=1e-10)
-    else:
-        result = rhs == lhs
-    return not result
-
-
-def _get_dict(entity):
-    """This removes all the entity attributes that are not considered part of
-    an AbstractCaomEntity comparison, and tracks the entities that are not
-    straight-to-dictionary conversions for later handling."""
-
-    attributes = {}
-    caom_collections = {}
-    for i in dir(entity):
-        attribute = getattr(entity, i)
-        if (i.startswith('_') or
-                callable(attribute) or
-                i.find('checksum') != -1):
-            continue
-        if (isinstance(attribute, TypedOrderedDict) or
-                isinstance(attribute, TypedList) or
-                isinstance(attribute, TypedSet) or
-                isinstance(attribute, CaomObject)):
-            caom_collections[i] = attribute
-        else:
-            attributes[i] = attribute
-    return attributes, caom_collections
 
 
 class TypedList(collections.MutableSequence):
